move ErrorCategory to types
