+++ packages/gatsby-cli/src/reporter/redux/internal-actions.ts
@@ -29,8 +29,7 @@ import {
   getElapsedTimeMS,
   getGlobalStatus,
 } from "./utils"
-import { IStructuredError } from "../../structured-errors/types"
-import { ErrorCategory } from "../../structured-errors/error-map"
+import { IStructuredError, ErrorCategory } from "../../structured-errors/types"
 import { IRenderPageArgs } from "../types"
 
 const ActivityStatusToLogLevel = {
